fixed tests splitting
