+++ papiea-engine/__tests__/entity_tests/extension.test.ts
@@ -61,7 +61,7 @@ describe("Entity API with metadata extension tests", () => {
         } catch (err) {
             const res = err.response;
             expect(res.status).toEqual(400);
-            expect(res.data.errors.length).toEqual(1);
+            expect(res.data.error.errors.length).toEqual(1);
 
         }
     });
@@ -111,7 +111,7 @@ describe("Entity API with metadata extension tests", () => {
         } catch (err) {
             const res = err.response;
             expect(res.status).toEqual(400);
-            expect(res.data.errors.length).toEqual(1);
+            expect(res.data.error.errors.length).toEqual(1);
         }
     });
 
@@ -138,8 +138,8 @@ describe("Entity API with metadata extension tests", () => {
         } catch (err) {
             const res = err.response;
             expect(res.status).toEqual(400);
-            expect(res.data.errors.length).toEqual(1);
-            expect(res.data.errors[0]).toEqual("Metadata extension is not specified");
+            expect(res.data.error.errors.length).toEqual(1);
+            expect(res.data.error.errors[0].message).toEqual("Metadata extension is not specified");
         }
     });
 });
